feat: Support overrides (#6)
If you need to perform a change that is technically breaking but still
permit it, `cue-schema` now permits multiple old and/or new schemas to
be supplied, and all schemas of a given type will be unified together.

This provides some partial ability to override breaking changes. For
example typically removing / redefining an enum field would be invalid:

Old:

```cue
#schema: enums: enum1: { value1?: 1, value2?: 2}
```

New:

```cue
#schema: enums: enum1: { value1?: 1 }
```

However by supplying an extra old schema, this can be made valid:

```cue
// mark value2 as bottom (forbidden)
#schema: enums: enum1: { value2: _|_ }
```